Move code from Gui to Duke for refactoring and create MainWindow Class
